pyproject.toml
[project]
name = "linkdin-scrapper"
version = "0.1.0"
description = ""
authors = [
    {name = "Akash",email = "akashkallai27@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "selenium (>=4.29.0,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
